style: remove unused useClassName and add test for useInjectStylesheet
